backend/productms/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
atomicwrites==1.4.1
attrs==22.1.0
click==8.1.3
colorama==0.4.6
coverage==6.5.0
Flask==2.2.2
flask-cors==3.0.10
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
SQLAlchemy-serializer==1.4.1
SQLAlchemy-Utils==0.38.3
toml==0.10.2
tomli==2.0.1
Werkzeug==2.2.2
apscheduler==3.9.1
Flask-Mail==0.9.1
Flask-Caching==2.0.1
redis==4.3.5
psycopg2==2.9.5
python-dotenv==0.21.0
flake8==6.0.0
pytest-mock==3.10.0